Add a lock_ttl to the sidekiq jobs
It's possible that the lock added by sidekiq-unique-jobs is not removed in exceptional scenarios (outage, server reboot, etc.). This prevents any further jobs from processing forever (or at least until the lock is cleared by redis expiry or dev intervention). Add a TTL so that in this case, things eventually start processing again. TTL is set to 1 hour, which seems like ample time to run jobs normally, but short enough that updates start working again quickly in the event of a problem.